src/requirements.txt
py_solc_x==1.1.1
pycryptodome==3.17
Requests==2.32.3
slither-analyzer==0.10.4
web3==6.20.3